+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### CURRENT
+* Added ability to open folder for download
 * Fixed Linux shell script mistake
 * Fixed issue on Linux system with main icon being too large